Aula 02 Banco de dados.

+++ Banco de Dados/Aula 01/SistemaBancario/SistemaBancario/Classes/Entidades/Banco.cs
@@ -10,7 +10,7 @@
         ///<sumary>
         /// Taxa fixa cobrada em cada operação de saque
         /// </sumary>
-        private const double taxaSaque = 5.00;
+        private const decimal taxaSaque = 5.00m;
 
         //Propriedades
         ///<summary>
@@ -38,7 +38,7 @@
         ///'private set' impede alteração direta - msó pode mudar através de Deposito ou Saque
         ///</summary>
         
-        public double Saldo { get; private set; }
+        public decimal Saldo { get; private set; }
 
         //Construtores
         /// <summary>
@@ -59,7 +59,7 @@
         ///  <param name="saldo">Valor do depósito inicial (opcional, padrão = 0)</param>
         
 
-        public Banco(int numeroConta, string titular, double saldo = 0)
+        public Banco(int numeroConta, string titular, decimal saldo = 0)
         {
             NumeroConta = numeroConta;
             Titular = titular;
@@ -72,7 +72,7 @@
         /// </summary>
         /// <param name="valor">Valor a ser depositado, deve ser positivo</param>
         /// 
-        public void Deposito(double valor)
+        public void Deposito(decimal valor)
         {
             if (valor<=0)
             {
@@ -94,7 +94,7 @@
         /// </summary>
         /// <param name="valor">Valor a ser sacado (deve ser positivo, não inclui a taxa)</param>
         /// 
-        public void Saque(double valor)
+        public void Saque(decimal valor)
         {
             if (valor <= 0)
             {
